Fix compiler warnings, add macros

+++ labs/lab-3/serial.ino
@@ -22,7 +22,7 @@ Command parseSerialInput(const char* input) {
 
 void readSerial(Stream& serialPort) {
   static char input[10];
-  static int index = 0;
+  static size_t index = 0;
 
   while (serialPort.available()) {
     char c = serialPort.read();
